Day 6/requirements.txt
fastapi
uvicorn
requests
python-dotenv
python-multipart 
assemblyai